Lógica de atualizar Produto agora contém Modal funcionando corretamente.
