+++ frontend/src/pages/admin/ProdutosAdmin/ProdutosAdmin.jsx
@@ -43,14 +43,6 @@ export const ProdutosAdmin = () => {
     setOpenModalProdutoCadastrado(true)
   }
 
-  // HANDLES DO MODAL DE ATUALIZAR
-  const handleCloseModalProdutoAtualizado = () => {
-    setOpenModalProdutoAtualizado(false)
-  }
-
-  const handleOpenModalProdutoAtualizado = () => {
-    setOpenModalProdutoAtualizado(true)
-  }
 
   // HANDLES DO MODAL DE EXCLUIR WARNING
   const handleCloseModalProdutoExcluirWarning = () => {
@@ -76,7 +68,7 @@ export const ProdutosAdmin = () => {
   }
 
   const [openModalProdutoCadastrado, setOpenModalProdutoCadastrado] = useState(false)
-  const [openModalProdutoAtualizado, setOpenModalProdutoAtualizado] = useState(false)
+
   const [openModalProdutoExcluirWarning, setOpenModalProdutoExcluirWarning] = useState(false)
   const [openModalProdutoExcluir, setOpenModalProdutoExcluir] = useState(false)
 
@@ -93,8 +85,6 @@ export const ProdutosAdmin = () => {
             {/* AO CONSTRUIR O FORMULARIO, COMENTAR A DIV ABAIXO, ESTOU USANDO ESTE BOTÃO PARA APLICAR OS MODALS */}
             <div>
               <div>
-
-                <input type="button" onClick={handleOpenModalProdutoAtualizado} value="Atualizar" />
                 <input type="button" onClick={handleOpenModalProdutoExcluirWarning} value="Excluir" />
               </div>
 
@@ -112,19 +102,7 @@ export const ProdutosAdmin = () => {
                 </div>
               </Modal>
 
-              {/* MODAL ATUALIZAR */}
-              <Modal isOpen={openModalProdutoAtualizado} onClick={handleCloseModalProdutoAtualizado}>
-                <div className={styles.conteudoModal}>
-                  <FontAwesomeIcon icon={faCircleCheck} className={styles.iconeModal} />
-                  <p> (nomeDoProduto) foi atualizado com sucesso! (puxar do backend)</p>
-                  <input
-                    className={styles.botaoConfirmarModal}
-                    type="button"
-                    value="OK"
-                    onClick={handleCloseModalProdutoAtualizado}
-                  />
-                </div>
-              </Modal>
+
 
               {/* MODAL EXCLUIR_WARNING */}
               <Modal isOpen={openModalProdutoExcluirWarning} onClick={handleCloseModalProdutoExcluirWarning}>
